requirements.txt
langchain==0.0.173
python-dotenv==1.0.0
streamlit==1.22.0
streamlit_chat==0.0.2.2

